Remove job_service_connector.php, repeat creating jobService in each sample
