+++ jobs/src/auto_complete_sample.php
@@ -17,6 +17,8 @@
 
 namespace Google\Cloud\Samples\Jobs;
 
+use Google_Client;
+use Google_Service_JobService;
 use Google_Service_JobService_CompleteQueryResponse;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
@@ -31,7 +33,42 @@ use Symfony\Component\Console\Output\OutputInterface;
  */
 final class AutoCompleteSample extends Command
 {
+    private static $jobService;
+
+    /**
+     * Creates a Google_Service_JobService with default application credentials.
+     * @return Google_Service_JobService
+     */
+    private static function create_job_service()
+    {
+        // Instantiate the client
+        $client = new Google_Client();
+
+        // Authorize the client using Application Default Credentials
+        // @see https://developers.google.com/identity/protocols/application-default-credentials
+        $client->useApplicationDefaultCredentials();
+        $client->setScopes(array('https://www.googleapis.com/auth/jobs'));
+
+        // Instantiate the Cloud Job Discovery Service API
+        $jobService = new Google_Service_JobService($client);
+        return $jobService;
+    }
+
+    /**
+     * Gets Google_Service_JobService.
+     *
+     * @return Google_Service_JobService
+     */
+    private static function get_job_service()
+    {
+        if (!isset(self::$jobService)) {
+            self::$jobService = self::create_job_service();
+        }
+        return self::$jobService;
+    }
+
     # [START auto_complete_job_title]
+
     /**
      * Auto completes job titles within given companyName.
      *
@@ -50,7 +87,7 @@ final class AutoCompleteSample extends Command
             $optParams['companyName'] = $companyName;
         }
 
-        $jobService = JobServiceConnector::get_job_service();
+        $jobService = self::get_job_service();
         $results = $jobService->v2->complete($optParams);
 
         var_export($results);
@@ -77,7 +114,7 @@ final class AutoCompleteSample extends Command
             $optParams['companyName'] = $companyName;
         }
 
-        $jobService = JobServiceConnector::get_job_service();
+        $jobService = self::get_job_service();
         $results = $jobService->v2->complete($optParams);
 
         var_export($results);
