+++ jobs/src/basic_company_sample.php
@@ -17,6 +17,8 @@
 
 namespace Google\Cloud\Samples\Jobs;
 
+use Google_Client;
+use Google_Service_JobService;
 use Google_Service_JobService_Company;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
@@ -39,8 +41,42 @@ use Symfony\Component\Console\Output\OutputInterface;
  */
 final class BasicCompanySample extends Command
 {
+    private static $jobService;
+
+    /**
+     * Creates a Google_Service_JobService with default application credentials.
+     * @return Google_Service_JobService
+     */
+    private static function create_job_service()
+    {
+        // Instantiate the client
+        $client = new Google_Client();
+
+        // Authorize the client using Application Default Credentials
+        // @see https://developers.google.com/identity/protocols/application-default-credentials
+        $client->useApplicationDefaultCredentials();
+        $client->setScopes(array('https://www.googleapis.com/auth/jobs'));
+
+        // Instantiate the Cloud Job Discovery Service API
+        $jobService = new Google_Service_JobService($client);
+        return $jobService;
+    }
+
+    /**
+     * Gets Google_Service_JobService.
+     *
+     * @return Google_Service_JobService
+     */
+    private static function get_job_service()
+    {
+        if (!isset(self::$jobService)) {
+            self::$jobService = self::create_job_service();
+        }
+        return self::$jobService;
+    }
 
     # [START basic_company]
+
     /**
      * Generates a company.
      *
@@ -69,7 +105,7 @@ final class BasicCompanySample extends Command
      */
     public static function create_company(Google_Service_JobService_Company $companyToBeCreated)
     {
-        $jobService = JobServiceConnector::get_job_service();
+        $jobService = self::get_job_service();
 
         $companyCreated = $jobService->companies->create($companyToBeCreated);
         printf("Company created:\n%s\n", var_export($companyCreated, true));
@@ -86,7 +122,7 @@ final class BasicCompanySample extends Command
      */
     public static function get_company(string $companyName)
     {
-        $jobService = JobServiceConnector::get_job_service();
+        $jobService = self::get_job_service();
 
         $companyExisted = $jobService->companies->get($companyName);
         printf("Company existed:\n%s\n", var_export($companyExisted, true));
@@ -104,7 +140,7 @@ final class BasicCompanySample extends Command
      */
     public static function update_company(string $companyName, Google_Service_JobService_Company $companyToBeUpdated)
     {
-        $jobService = JobServiceConnector::get_job_service();
+        $jobService = self::get_job_service();
 
         $companyUpdated = $jobService->companies->patch($companyName, $companyToBeUpdated);
         printf("Company updated:\n%s\n", var_export($companyUpdated, true));
@@ -125,9 +161,8 @@ final class BasicCompanySample extends Command
         string $companyName,
         string $fieldMask,
         Google_Service_JobService_Company $companyToBeUpdated
-    )
-    {
-        $jobService = JobServiceConnector::get_job_service();
+    ) {
+        $jobService = self::get_job_service();
 
         $optParams = array('updateCompanyFields' => $fieldMask);
         $companyUpdated = $jobService->companies->patch($companyName, $companyToBeUpdated, $optParams);
@@ -144,7 +179,7 @@ final class BasicCompanySample extends Command
      */
     public static function delete_company(string $companyName)
     {
-        $jobService = JobServiceConnector::get_job_service();
+        $jobService = self::get_job_service();
 
         $jobService->companies->delete($companyName);
         echo 'Company deleted' . PHP_EOL;
