+++ jobs/src/batch_operation_sample.php
@@ -17,7 +17,9 @@
 
 namespace Google\Cloud\Samples\Jobs;
 
+use Google_Client;
 use Google_Service_Exception;
+use Google_Service_JobService;
 use Google_Service_JobService_CreateJobRequest;
 use Google_Service_JobService_Job;
 use Google_Service_JobService_UpdateJobRequest;
@@ -39,7 +41,42 @@ use Symfony\Component\Console\Output\OutputInterface;
  */
 final class BatchOperationSample extends Command
 {
+    private static $jobService;
+
+    /**
+     * Creates a Google_Service_JobService with default application credentials.
+     * @return Google_Service_JobService
+     */
+    private static function create_job_service()
+    {
+        // Instantiate the client
+        $client = new Google_Client();
+
+        // Authorize the client using Application Default Credentials
+        // @see https://developers.google.com/identity/protocols/application-default-credentials
+        $client->useApplicationDefaultCredentials();
+        $client->setScopes(array('https://www.googleapis.com/auth/jobs'));
+
+        // Instantiate the Cloud Job Discovery Service API
+        $jobService = new Google_Service_JobService($client);
+        return $jobService;
+    }
+
+    /**
+     * Gets Google_Service_JobService.
+     *
+     * @return Google_Service_JobService
+     */
+    private static function get_job_service()
+    {
+        if (!isset(self::$jobService)) {
+            self::$jobService = self::create_job_service();
+        }
+        return self::$jobService;
+    }
+
     # [START batch_job_create]
+
     /**
      * Creates jobs in batch.
      *
@@ -48,7 +85,7 @@ final class BatchOperationSample extends Command
      */
     public static function batch_create_jobs(string $companyName)
     {
-        $jobService = JobServiceConnector::get_job_service();
+        $jobService = self::get_job_service();
         $jobService->getClient()->setUseBatch(true);
 
         $softwareEngineerJob = new Google_Service_JobService_Job();
@@ -105,7 +142,7 @@ final class BatchOperationSample extends Command
      */
     public static function batch_job_update(Google_Service_JobService_Job ...$jobsToBeUpdated)
     {
-        $jobService = JobServiceConnector::get_job_service();
+        $jobService = self::get_job_service();
         $jobService->getClient()->setUseBatch(true);
 
         // Creates batch request.
@@ -156,7 +193,7 @@ final class BatchOperationSample extends Command
      */
     public static function batch_delete_jobs(Google_Service_JobService_Job ...$jobsToBeDeleted)
     {
-        $jobService = JobServiceConnector::get_job_service();
+        $jobService = self::get_job_service();
         $jobService->getClient()->setUseBatch(true);
 
         // Creates batch request.
