+++ jobs/src/commute_search_sample.php
@@ -17,6 +17,8 @@
 
 namespace Google\Cloud\Samples\Jobs;
 
+use Google_Client;
+use Google_Service_JobService;
 use Google_Service_JobService_CommutePreference;
 use Google_Service_JobService_JobQuery;
 use Google_Service_JobService_LatLng;
@@ -35,7 +37,42 @@ use Symfony\Component\Console\Output\OutputInterface;
  */
 final class CommuteSearchSample extends Command
 {
+    private static $jobService;
+
+    /**
+     * Creates a Google_Service_JobService with default application credentials.
+     * @return Google_Service_JobService
+     */
+    private static function create_job_service()
+    {
+        // Instantiate the client
+        $client = new Google_Client();
+
+        // Authorize the client using Application Default Credentials
+        // @see https://developers.google.com/identity/protocols/application-default-credentials
+        $client->useApplicationDefaultCredentials();
+        $client->setScopes(array('https://www.googleapis.com/auth/jobs'));
+
+        // Instantiate the Cloud Job Discovery Service API
+        $jobService = new Google_Service_JobService($client);
+        return $jobService;
+    }
+
+    /**
+     * Gets Google_Service_JobService.
+     *
+     * @return Google_Service_JobService
+     */
+    private static function get_job_service()
+    {
+        if (!isset(self::$jobService)) {
+            self::$jobService = self::create_job_service();
+        }
+        return self::$jobService;
+    }
+
     # [START commute_search]
+
     /**
      * Search jobs based on commute location and time.
      *
@@ -75,12 +112,14 @@ final class CommuteSearchSample extends Command
         $searchRequest->setJobView('FULL');
         $searchRequest->setEnablePreciseResultSize(true);
 
-        $jobService = JobServiceConnector::get_job_service();
+        $jobService = self::get_job_service();
         $response = $jobService->jobs->search($searchRequest);
 
         var_export($response);
         return $response;
     }
+
+    # [END commute_search]
 
     protected function configure()
     {
