+++ jobs/src/custom_attribute_sample.php
@@ -17,7 +17,8 @@
 
 namespace Google\Cloud\Samples\Jobs;
 
-
+use Google_Client;
+use Google_Service_JobService;
 use Google_Service_JobService_CustomAttribute;
 use Google_Service_JobService_Job;
 use Google_Service_JobService_JobQuery;
@@ -38,7 +39,42 @@ use Symfony\Component\Console\Output\OutputInterface;
  */
 final class CustomAttributeSample extends Command
 {
+    private static $jobService;
+
+    /**
+     * Creates a Google_Service_JobService with default application credentials.
+     * @return Google_Service_JobService
+     */
+    private static function create_job_service()
+    {
+        // Instantiate the client
+        $client = new Google_Client();
+
+        // Authorize the client using Application Default Credentials
+        // @see https://developers.google.com/identity/protocols/application-default-credentials
+        $client->useApplicationDefaultCredentials();
+        $client->setScopes(array('https://www.googleapis.com/auth/jobs'));
+
+        // Instantiate the Cloud Job Discovery Service API
+        $jobService = new Google_Service_JobService($client);
+        return $jobService;
+    }
+
+    /**
+     * Gets Google_Service_JobService.
+     *
+     * @return Google_Service_JobService
+     */
+    private static function get_job_service()
+    {
+        if (!isset(self::$jobService)) {
+            self::$jobService = self::create_job_service();
+        }
+        return self::$jobService;
+    }
+
     # [START custom_attribute_job]
+
     /**
      * Generates a job with a custom attribute.
      *
@@ -100,7 +136,7 @@ final class CustomAttributeSample extends Command
         $searchJobsRequest->setRequestMetadata($requestMetadata);
         $searchJobsRequest->setJobView('FULL');
 
-        $response = JobServiceConnector::get_job_service()->jobs->search($searchJobsRequest);
+        $response = self::get_job_service()->jobs->search($searchJobsRequest);
         var_export($response);
         return $response;
     }
@@ -132,7 +168,7 @@ final class CustomAttributeSample extends Command
         $searchJobsRequest->setRequestMetadata($requestMetadata);
         $searchJobsRequest->setJobView('FULL');
 
-        $response = JobServiceConnector::get_job_service()->jobs->search($searchJobsRequest);
+        $response = self::get_job_service()->jobs->search($searchJobsRequest);
         var_export($response);
         return $response;
     }
@@ -164,7 +200,7 @@ final class CustomAttributeSample extends Command
         $searchJobsRequest->setRequestMetadata($requestMetadata);
         $searchJobsRequest->setJobView('FULL');
 
-        $response = JobServiceConnector::get_job_service()->jobs->search($searchJobsRequest);
+        $response = self::get_job_service()->jobs->search($searchJobsRequest);
         var_export($response);
         return $response;
     }
