+++ jobs/src/general_search_sample.php
@@ -17,6 +17,8 @@
 
 namespace Google\Cloud\Samples\Jobs;
 
+use Google_Client;
+use Google_Service_JobService;
 use Google_Service_JobService_CompensationEntry;
 use Google_Service_JobService_CompensationFilter;
 use Google_Service_JobService_CompensationInfo;
@@ -49,7 +51,41 @@ use Symfony\Component\Console\Output\OutputInterface;
  */
 final class GeneralSearchSample extends Command
 {
+    private static $jobService;
+
+    /**
+     * Creates a Google_Service_JobService with default application credentials.
+     * @return Google_Service_JobService
+     */
+    private static function create_job_service()
+    {
+        // Instantiate the client
+        $client = new Google_Client();
+
+        // Authorize the client using Application Default Credentials
+        // @see https://developers.google.com/identity/protocols/application-default-credentials
+        $client->useApplicationDefaultCredentials();
+        $client->setScopes(array('https://www.googleapis.com/auth/jobs'));
+
+        // Instantiate the Cloud Job Discovery Service API
+        $jobService = new Google_Service_JobService($client);
+        return $jobService;
+    }
+
+    /**
+     * Gets Google_Service_JobService.
+     *
+     * @return Google_Service_JobService
+     */
+    private static function get_job_service()
+    {
+        if (!isset(self::$jobService)) {
+            self::$jobService = self::create_job_service();
+        }
+        return self::$jobService;
+    }
     # [START basic_keyword_search]
+
     /**
      * Simple job search with keyword.
      *
@@ -80,7 +116,7 @@ final class GeneralSearchSample extends Command
         $searchRequest->setQuery($jobQuery);
         $searchRequest->setMode('JOB_SEARCH');
 
-        $jobService = JobServiceConnector::get_job_service();
+        $jobService = self::get_job_service();
         $response = $jobService->jobs->search($searchRequest);
 
         var_export($response);
@@ -118,7 +154,7 @@ final class GeneralSearchSample extends Command
         $searchRequest->setQuery($jobQuery);
         $searchRequest->setMode('JOB_SEARCH');
 
-        $jobService = JobServiceConnector::get_job_service();
+        $jobService = self::get_job_service();
         $response = $jobService->jobs->search($searchRequest);
 
         var_export($response);
@@ -156,7 +192,7 @@ final class GeneralSearchSample extends Command
         $searchRequest->setQuery($jobQuery);
         $searchRequest->setMode('JOB_SEARCH');
 
-        $jobService = JobServiceConnector::get_job_service();
+        $jobService = self::get_job_service();
         $response = $jobService->jobs->search($searchRequest);
 
         var_export($response);
@@ -194,7 +230,7 @@ final class GeneralSearchSample extends Command
         $searchRequest->setQuery($jobQuery);
         $searchRequest->setMode('JOB_SEARCH');
 
-        $jobService = JobServiceConnector::get_job_service();
+        $jobService = self::get_job_service();
         $response = $jobService->jobs->search($searchRequest);
 
         var_export($response);
@@ -232,7 +268,7 @@ final class GeneralSearchSample extends Command
         $searchRequest->setQuery($jobQuery);
         $searchRequest->setMode('JOB_SEARCH');
 
-        $jobService = JobServiceConnector::get_job_service();
+        $jobService = self::get_job_service();
         $response = $jobService->jobs->search($searchRequest);
 
         var_export($response);
@@ -270,7 +306,7 @@ final class GeneralSearchSample extends Command
         $searchRequest->setQuery($jobQuery);
         $searchRequest->setMode('JOB_SEARCH');
 
-        $jobService = JobServiceConnector::get_job_service();
+        $jobService = self::get_job_service();
         $response = $jobService->jobs->search($searchRequest);
 
         var_export($response);
@@ -323,7 +359,7 @@ final class GeneralSearchSample extends Command
         $searchRequest->setQuery($jobQuery);
         $searchRequest->setMode('JOB_SEARCH');
 
-        $jobService = JobServiceConnector::get_job_service();
+        $jobService = self::get_job_service();
         $response = $jobService->jobs->search($searchRequest);
 
         var_export($response);
