+++ jobs/src/histogram_sample.php
@@ -17,7 +17,8 @@
 
 namespace Google\Cloud\Samples\Jobs;
 
-
+use Google_Client;
+use Google_Service_JobService;
 use Google_Service_JobService_CustomAttributeHistogramRequest;
 use Google_Service_JobService_HistogramFacets;
 use Google_Service_JobService_JobQuery;
@@ -33,7 +34,42 @@ use Symfony\Component\Console\Output\OutputInterface;
  */
 final class HistogramSample extends Command
 {
+    private static $jobService;
+
+    /**
+     * Creates a Google_Service_JobService with default application credentials.
+     * @return Google_Service_JobService
+     */
+    private static function create_job_service()
+    {
+        // Instantiate the client
+        $client = new Google_Client();
+
+        // Authorize the client using Application Default Credentials
+        // @see https://developers.google.com/identity/protocols/application-default-credentials
+        $client->useApplicationDefaultCredentials();
+        $client->setScopes(array('https://www.googleapis.com/auth/jobs'));
+
+        // Instantiate the Cloud Job Discovery Service API
+        $jobService = new Google_Service_JobService($client);
+        return $jobService;
+    }
+
+    /**
+     * Gets Google_Service_JobService.
+     *
+     * @return Google_Service_JobService
+     */
+    private static function get_job_service()
+    {
+        if (!isset(self::$jobService)) {
+            self::$jobService = self::create_job_service();
+        }
+        return self::$jobService;
+    }
+
     # [START histogram_search]
+
     /**
      * Histogram search.
      *
@@ -70,7 +106,7 @@ final class HistogramSample extends Command
             $request->setQuery($jobQuery);
         }
 
-        $response = JobServiceConnector::get_job_service()->jobs->search($request);
+        $response = self::get_job_service()->jobs->search($request);
         var_export($response);
         return $response;
     }
