+++ jobs/test/CommuteSearchSampleTest.php
@@ -17,7 +17,6 @@
 
 namespace Google\Cloud\Samples\Jobs;
 
-
 use Symfony\Component\Console\Tester\CommandTester;
 
 class CommuteSearchSampleTest extends \PHPUnit_Framework_TestCase
